fix(website): Removed Switching Camera
Cross-compatibility was a mess, it only worked part of the time, and at the end of the day people aren't going to be taking a picture of their face with a plant next to it. The assumption is made that the environment camera will now be the only one used.