fix(website): Fixed Cross-Compatibility Issue to Support Mobile Devices (#9)
* fix(website): Added More Descriptive Error Logging for Camera Errors

* fix(website): Switched From DeviceId to FacingMode for Camera Selection

* fix(website): Added Support for Devices that Don't Support FacingMode

* fix(website): Fixed Camera Orientation Issue on Android

* fix(website): Fix Camera Not Reading on Safari

* fix(website): Removed Switching Camera

Cross-compatibility was a mess, it only worked part of the time, and at the end of the day people aren't going to be taking a picture of their face with a plant next to it. The assumption is made that the environment camera will now be the only one used.